Try performs multiple attempt to delete compile-transaction
In between the attempts, we perform a JVM Garbage Collection as it may
help. There are several strategies we could use such as deferring the
deletes until the build finishes or the JVM exits or ignoring the
failure completely. It seems like the FileSystemOperations#delete and
Deleter#deleteRecursively doesn't offer enough flexibility. However, the
more flexibility we offer, the more complex the deletion will become. On
the other hand, the more smart we include, the more performance
degradation we may see. We could leverage the WorkAction to delete in
the background. Let's try with Files.move and GC retry.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>